requirements.txt
fastapi
uvicorn
requests
pyjwt
python-telegram-bot
sqlalchemy
